inventory: Fix Inventory Usage And Plugin Variables
Assigning to plugin native ansible variables.

This playbook uses an infrared role that is responsible
for generating an inventory inside the workspace.
The generated infrared inventory is different than our generated
inventory due to a different Jinja2 template.
Most common issue is when invoking 'tripleo-overcloud' infrared plugin
after 'tripleo-clouds-inventory' an ansible group '[overcloud_nodes]'
is missing and tripleo-overcloud is failing.

This patch adding a workaround for 'inventory-update' role that
adding an empty 'overcloud_nodes' group to ansible in memory inventory
which is used for generating inventory for infrared workspace.

Change-Id: If68b8dd703d871f254f3cd4e8567ddfc7b911f30